fix: fix overwrite for td_bulk_upload

+++ R/td.R
@@ -95,11 +95,11 @@ td_embulk_upload <- function(conn, dbname, table, df, embulk_dir, overwrite = FA
 td_bulk_upload <- function(conn, dbname, table, df, overwrite = FALSE, append = FALSE){
   # Check if msgpack package installed
   check_msgpack()
-  if (overwrite) {
+  exists_table <- exist_table(conn, dbname, table)
+  if (exists_table && overwrite) {
     delete_table(conn, dbname, table)
     create_table(conn, dbname, table)
   }
-  exists_table <- exist_table(conn, dbname, table)
   if (!exists_table) {
     create_table(conn, dbname, table)
   }
